Add support for fluentd (EFK)
This adds support for using fluentd as a replacement for logstash.
You'll need to use the new logging_backend option in
groupvars/all.yml to use this (supports rsyslog only).

Logstash and filebeat are still the defaults.  Also making firewall
and service ports configurable as variables so things are more
flexible for differing deployments.