Refactored Beats into same ELK-Client configuration
Added packetbeat, metricbeat, and heartbeat configuration back to the elk_client standard. Created the appropriate variables and separated them out. I moved the installation of the Elastic.co repo to the role ELK_Client so as to not have to maintain 4 repositories when they are duplicates. I also added the correct installation variables to all.yml and added the installation conditions to elk_client and elk.