Force manual update of logstash-input-udp.
There's a bug with the current logstash-input-udp plugin that
ships with the latest Elastic.co logstash rpm packages.

This ships with 3.3.1 and it's fixed in 3.3.2, so let's force
a manual update of this specific plugin via:

/usr/share/logstash/bin/logstash-plugin update logstash-input-udp

In the future I might consider updating all plugins to latest if
this happens again.

Fixes: https://github.com/sadsfae/ansible-elk/issues/69