5.6 Series is now moved to a maintained branch.
We will be moving to the 6.x ELK series and master branch
will become 6.x series.  What is currently master will become
the 5.6 branch similar to what we've done for 2.4.

Fixes:

I've made some changes to yum with_items as this will be deprecated
in Ansible 2.11+.
Curator should be working again for 5.6, thanks @crossan007
for filing an issue about it.

Fixes: https://github.com/sadsfae/ansible-elk/issues/71
Related-to: https://github.com/sadsfae/ansible-elk/issues/66