Support Apache as Reverse Proxy + Fixes.
This implements an option for Apache to be used as a reverse proxy.
This also updates the code to remove deprecation warnings from newer
versions of ansible.

Details:

In install/group_vars/all.yml

apache_reverse_proxy: true

The default will continue to be nginx.

Note that this change requires setting the kibana basepath in order to
use Location based ProxyPass functionality in Apache so we'll manage
kibana.yml now via jinja2 template.

Fixes: https://github.com/sadsfae/ansible-elk/issues/77