Fixes for heartbeat package/config changes, TLS
There are changes to the heartbeat service that need changing for it
work correctly.

Further, we're going to remove SSL/TLS certificates and keys that exist
for elk_client playbooks runs to ensure we have the latest ones
installed in case they change from the elk server.

Note that if you're using heartbeat you'll need to configure
heartbeat.yml yourself further for other monitoring outside logstash.

fixes: https://github.com/sadsfae/ansible-elk/issues/79
related-to: https://github.com/sadsfae/ansible-elk/issues/76