Add handler for filebeat service.
Since we're deploying new TLS/SSL certificates (even if they haven't
changed - same ones) it's a good idea to associate a filebeat service
restart handler when that happens.

This will catch cases where a new TLS/SSL certificate is deployed and
the service might not be restarted, and there's no harm in doing it
otherwise.